Queue a PR refresh if CI status is pending
This is a rebase of https://github.com/voxpupuli/vox-pupuli-tasks/pull/214
The original commit message:

---

the status is pending if one configured check is still running. If a PR
was just created, the checks haven't passed/failed yet. This wouldn't be
bad IF we would get notifications if they finish. We could implement
travis CI notification parsing for this, but people might implement
other checks as well. We can simply queue a job that updates the PR.

This fixes #213 in parts. If we update the PR and the status is still
pending, we don't queue another job. After some thoughts it's probably a
good idea to move the logic from validate_status() to
`queue_validation()` and `validate()`.