Improve Operating System support stats
All modules do not support all operating systems.

The #check_operating_system_support method iterates on each supported
operating system and set `supports_only_current_<os>` and
`supports_latest_<os>` accordingly.  We can therefore check if a value
was registered or not for them, i.e. true and false vs. nil.

Each OS will then display the number of up-to-date modules compared to
the number of modules which idealy should be up-to-date, and not
compared to all existing modules.